config-service/pyproject.toml
[project]
name = "config-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.116.1",
    "psycopg2==2.9.10",
    "pydantic==2.11.7",
    "pydantic-extra-types>=2.11.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "python-ulid>=2.0.0,<3.0.0",
    "uvicorn>=0.40.0",
    "httpx>=0.28.0",
    "PyJWT>=2.8.0",
]

[dependency-groups]
dev = [
    "httpx==0.28.1",
    "pytest==8.4.1",
    "pytest-asyncio>=1.3.0",
]
